pyproject.toml
[tool.poetry]
name = "cli-pytest-async-lab"
version = "0.1.0"
description = ""
authors = ["Falk <gammalget@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
black = "^21.8b0"
isort = "^5.9.3"
pytest-asyncio = "^0.15.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
